Updated home map class

+++ MelbourneSight/Sight.swift
@@ -11,12 +11,12 @@ import UIKit
 class Sight: NSObject {
     var name: String
     var desc: String
-    var latitude: Float
-    var longitude: Float
+    var latitude: Double
+    var longitude: Double
     var mapIcon: String
     var photo: String
     
-    init(name: String, desc: String, latitude: Float, longitude: Float, mapIcon: String, photo: String) {
+    init(name: String, desc: String, latitude: Double, longitude: Double, mapIcon: String, photo: String) {
         self.name = name
         self.desc = desc
         self.latitude = latitude
